Move get_transaction_amount to external_api.py. Edit test, now all tests working!
