+++ utils/external_api.py
@@ -18,3 +18,13 @@ def convert_to_rub(amount: float, _form: str) -> float:
 
     result = json.loads(response.text)['result']
     return result
+
+
+def get_transaction_amount(transaction: dict) -> float:
+    """Return transaction amount in RUB"""
+    amount = float(transaction['operationAmount']['amount'])
+    currency_code = transaction['operationAmount']['currency']["code"]
+    if currency_code == "RUB":
+        return amount
+    amount = float(convert_to_rub(amount, currency_code))
+    return amount
